Created branch `local`, edited config.h and config.mk
The `local` branch is the branch where my edits and configurations for DWM will be located from now on.
The original config.mk needs to be modified to be used on FreeBSD because of library location differences.
The config.h file is made from config.def.h. From now on, edits will be made to config.h file, not the config.def.h file.