Changed dmenucmd spawn keybind from XK_p to XK_r
I want to make it more "semantically correct" as to what dmenu's usage is.
So, dmenu is used to launch / `run` things, so it will be more informative if I map
the keybind for dmenu to XK_r ('r' meaning "run", just like Windows' Win-R).
I would like to think of XK_p as "prompt". It's likely that the 'p' there means "prompt".
But, I want to map `shiftview` (the patch to enable tag switching like ALT-Tab in Windows or Command-Tab in MacOS) to XK_p for "previous tag" and XK_n for "next tag".